+++ src/redux/store.js
@@ -1,6 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
 import rocketsReducer from './rocketsSlice';
+import missionReducer from './missionSlice';
 
 const store = configureStore({
   reducer: {
